refactor(indexer): extract cartog-indexer test module into focused files
The ~2460-line #[cfg(test)] mod tests (65 tests) moves out of lib.rs into
src/tests/{jobs,walk_filter,summary,incremental,git,dedup,merkle,progress,
redaction}.rs. lib.rs drops 3744 -> 1280 lines; production code byte-identical.
Tests stay in-crate (private access via crate::*); visible_dir hoisted to the
tests hub. No test added, dropped, or modified.
