+++ crates/cartog-indexer/src/lib.rs
@@ -1277,2468 +1277,4 @@ pub mod bench_support {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn clamp_jobs_resolves_zero_to_available_and_bounds_the_rest() {
-        // 0 = auto, but capped at 64 even on a >64-core host.
-        let auto = std::thread::available_parallelism()
-            .map(|p| p.get())
-            .unwrap_or(1)
-            .min(64);
-        assert_eq!(clamp_jobs(0), auto, "0 = auto, clamped to the upper bound");
-        assert_eq!(clamp_jobs(1), 1);
-        assert_eq!(clamp_jobs(8), 8);
-        assert_eq!(clamp_jobs(64), 64);
-        assert_eq!(clamp_jobs(1000), 64, "clamped to the upper bound");
-    }
-
-    #[test]
-    fn parse_pool_is_sized_to_jobs_and_reused() {
-        // A sized pool actually has that many threads, and the same size hands
-        // back the cached instance (warm threads survive re-index).
-        let p3 = parse_pool(3).expect("pool builds");
-        assert_eq!(p3.current_num_threads(), 3);
-        let p3_again = parse_pool(3).expect("pool builds");
-        assert!(
-            Arc::ptr_eq(&p3, &p3_again),
-            "same size reuses the cached pool"
-        );
-        let p1 = parse_pool(1).expect("pool builds");
-        assert_eq!(p1.current_num_threads(), 1);
-    }
-
-    fn index_dir(db: &cartog_db::Database, root: &Path, force: bool) -> Result<IndexResult> {
-        index_directory(
-            db,
-            root,
-            force,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-    }
-
-    fn index_dir_with_jobs(db: &cartog_db::Database, root: &Path, jobs: usize) -> IndexResult {
-        let filter = WalkFilter {
-            jobs,
-            ..Default::default()
-        };
-        index_directory(
-            db,
-            root,
-            false,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &filter,
-        )
-        .unwrap()
-    }
-
-    fn symbol_names(db: &cartog_db::Database) -> Vec<String> {
-        let files = db.all_files().unwrap();
-        db.symbols_for_files(&files, None)
-            .unwrap()
-            .into_iter()
-            .map(|s| s.name)
-            .collect()
-    }
-
-    #[test]
-    fn index_output_is_identical_across_pool_sizes() {
-        // The dedicated parse pool applies per call, so two pool sizes are
-        // comparable in one process: parallel parsing must not change results.
-        let make = || {
-            let tmp = tempfile::TempDir::new().unwrap();
-            let proj = visible_dir(&tmp, "proj");
-            for i in 0..12 {
-                std::fs::write(
-                    proj.join(format!("m{i}.py")),
-                    format!("def f{i}():\n    return {i}\n"),
-                )
-                .unwrap();
-            }
-            (tmp, proj)
-        };
-
-        let db1 = cartog_db::Database::open_memory().unwrap();
-        let (_t1, p1) = make();
-        let r1 = index_dir_with_jobs(&db1, &p1, 1);
-
-        let db8 = cartog_db::Database::open_memory().unwrap();
-        let (_t8, p8) = make();
-        let r8 = index_dir_with_jobs(&db8, &p8, 8);
-
-        assert_eq!(r1.files_indexed, 12);
-        assert_eq!(r1.files_indexed, r8.files_indexed);
-        assert_eq!(r1.symbols_added, r8.symbols_added);
-
-        let mut names1 = symbol_names(&db1);
-        let mut names8 = symbol_names(&db8);
-        names1.sort();
-        names8.sort();
-        assert_eq!(names1, names8, "symbol set is pool-size independent");
-    }
-
-    // TempDir roots are dot-prefixed (`.tmpXXXX`) and the walker skips hidden
-    // roots entirely — give each test a visible subdir to index.
-    fn visible_dir(tmp: &tempfile::TempDir, name: &str) -> std::path::PathBuf {
-        let dir = tmp.path().join(name);
-        std::fs::create_dir(&dir).unwrap();
-        dir
-    }
-
-    #[test]
-    fn index_refuses_empty_root_when_db_has_files() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let src = visible_dir(&tmp, "proj");
-        std::fs::write(src.join("a.py"), "def foo():\n    return 1\n").unwrap();
-        index_dir(&db, &src, false).unwrap();
-        assert!(!db.all_files().unwrap().is_empty());
-
-        // Pointing the same DB at a root with no supported files (the
-        // `rag index --db X` from-a-wrong-cwd footgun) must refuse instead
-        // of sweeping every indexed file away.
-        let empty = visible_dir(&tmp, "elsewhere");
-        let res = index_dir(&db, &empty, false);
-        assert!(res.is_err(), "expected refusal, got {res:?}");
-        assert!(
-            !db.all_files().unwrap().is_empty(),
-            "index must be left untouched on refusal"
-        );
-    }
-
-    #[test]
-    fn index_force_allows_emptying_the_index() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let src = visible_dir(&tmp, "proj");
-        std::fs::write(src.join("a.py"), "def foo():\n    return 1\n").unwrap();
-        index_dir(&db, &src, false).unwrap();
-
-        let empty = visible_dir(&tmp, "elsewhere");
-        index_dir(&db, &empty, true).unwrap();
-        assert!(db.all_files().unwrap().is_empty());
-    }
-
-    fn index_dir_excluding(
-        db: &cartog_db::Database,
-        root: &Path,
-        patterns: &[&str],
-    ) -> Result<IndexResult> {
-        let exclude = ExcludeGlobs::from_globs(
-            &patterns
-                .iter()
-                .map(|s| (*s).to_string())
-                .collect::<Vec<_>>(),
-        )
-        .unwrap();
-        let filter = WalkFilter {
-            exclude,
-            respect_gitignore: true,
-            ..Default::default()
-        };
-        index_directory(
-            db,
-            root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &filter,
-        )
-    }
-
-    #[test]
-    fn exclude_prunes_dir_and_matches_files() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::write(proj.join("a.py"), "def foo():\n    return 1\n").unwrap();
-        std::fs::create_dir(proj.join("Pods")).unwrap();
-        std::fs::write(proj.join("Pods/b.py"), "def bar():\n    return 2\n").unwrap();
-        // A non-source file inside the excluded dir: if the dir were merely
-        // file-filtered (walked + each file dropped) rather than PRUNED, the walk
-        // would tally this as unsupported. Pruning means it's never visited.
-        std::fs::write(proj.join("Pods/readme.xyz"), "junk\n").unwrap();
-        std::fs::write(proj.join("notes.md"), "# Title\n\nbody\n").unwrap();
-
-        let r = index_dir_excluding(&db, &proj, &["Pods/**", "**/*.md"]).unwrap();
-
-        assert_eq!(r.files_indexed, 1, "only a.py should be indexed");
-        assert_eq!(
-            r.files_unsupported, 0,
-            "Pods/ must be pruned (not descended), so its .xyz is never tallied"
-        );
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.ends_with("a.py")));
-        assert!(
-            !files.iter().any(|f| f.contains("Pods")),
-            "pruned dir must contribute no files"
-        );
-        assert!(!files.iter().any(|f| f.ends_with(".md")));
-    }
-
-    #[test]
-    fn exclude_matching_everything_errors_with_exclude_hint() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::write(proj.join("a.py"), "def foo():\n    return 1\n").unwrap();
-        index_dir(&db, &proj, false).unwrap();
-
-        // A valid glob that happens to match every source file empties the walk.
-        // Non-force must refuse AND point at the exclude, not just "wrong root".
-        let filter = WalkFilter {
-            exclude: ExcludeGlobs::from_globs(&["**/*.py".to_string()]).unwrap(),
-            respect_gitignore: true,
-            ..Default::default()
-        };
-        let res = index_directory(
-            &db,
-            &proj,
-            false,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &filter,
-        );
-        let err = format!("{:#}", res.unwrap_err());
-        assert!(err.contains("[index] exclude"), "hint missing: {err}");
-        assert!(!db.all_files().unwrap().is_empty(), "index left untouched");
-    }
-
-    #[test]
-    fn exclude_empty_is_noop() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::write(proj.join("a.py"), "def foo():\n    return 1\n").unwrap();
-        std::fs::create_dir(proj.join("Pods")).unwrap();
-        std::fs::write(proj.join("Pods/b.py"), "def bar():\n    return 2\n").unwrap();
-        std::fs::write(proj.join("notes.md"), "# Title\n\nbody\n").unwrap();
-
-        let r = index_dir_excluding(&db, &proj, &[]).unwrap();
-
-        // No globs → everything supported is indexed (both .py + the .md doc).
-        assert_eq!(r.files_indexed, 3);
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.contains("Pods")));
-        assert!(files.iter().any(|f| f.ends_with(".md")));
-    }
-
-    fn index_dir_filtered(
-        db: &cartog_db::Database,
-        root: &Path,
-        filter: &WalkFilter,
-    ) -> Result<IndexResult> {
-        index_directory(
-            db,
-            root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            filter,
-        )
-    }
-
-    #[test]
-    fn gitignore_nested_is_honored() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::create_dir_all(proj.join("sub/ignored")).unwrap();
-        // Nested .gitignore (not at the root) — e.g. a CocoaPods Pods/ dir.
-        std::fs::write(proj.join("sub/.gitignore"), "ignored/\n").unwrap();
-        std::fs::write(proj.join("sub/keep.py"), "def keep():\n    pass\n").unwrap();
-        std::fs::write(proj.join("sub/ignored/skip.py"), "def skip():\n    pass\n").unwrap();
-
-        let r = index_dir_filtered(&db, &proj, &WalkFilter::unrestricted()).unwrap();
-
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.ends_with("keep.py")));
-        assert!(
-            !files.iter().any(|f| f.contains("ignored")),
-            "nested .gitignore must hide sub/ignored/; got {files:?}"
-        );
-        assert_eq!(r.files_indexed, 1);
-    }
-
-    #[test]
-    fn gitignore_ancestor_above_root_is_not_applied() {
-        // A .gitignore ABOVE the indexed root (parent dir, or $HOME/.gitignore)
-        // must NOT prune files inside the root — else indexing a subdir of a
-        // repo silently drops files matched by the repo-root or $HOME ignore.
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        // Ignore file in the PARENT of what we index.
-        std::fs::write(tmp.path().join(".gitignore"), "*.log\nsecret/\n").unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::create_dir(proj.join("secret")).unwrap();
-        std::fs::write(proj.join("app.py"), "def app():\n    pass\n").unwrap();
-        std::fs::write(proj.join("debug.log"), "noise\n").unwrap();
-        std::fs::write(proj.join("secret/s.py"), "def s():\n    pass\n").unwrap();
-
-        index_dir_filtered(&db, &proj, &WalkFilter::unrestricted()).unwrap();
-
-        // The ancestor's `*.log`/`secret/` rules must not reach into proj/.
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.ends_with("app.py")));
-        assert!(
-            files.iter().any(|f| f.ends_with("secret/s.py")),
-            "ancestor .gitignore must NOT prune in-root secret/; got {files:?}"
-        );
-    }
-
-    #[test]
-    fn gitignore_floor_wins_over_unignore() {
-        // Even if .gitignore `!`-unignores node_modules, the hardcoded floor
-        // still skips it.
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::write(proj.join(".gitignore"), "!node_modules/\n").unwrap();
-        std::fs::create_dir(proj.join("node_modules")).unwrap();
-        std::fs::write(proj.join("node_modules/x.py"), "def x():\n    pass\n").unwrap();
-        std::fs::write(proj.join("app.py"), "def app():\n    pass\n").unwrap();
-
-        index_dir_filtered(&db, &proj, &WalkFilter::unrestricted()).unwrap();
-
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.ends_with("app.py")));
-        assert!(!files.iter().any(|f| f.contains("node_modules")));
-    }
-
-    #[test]
-    fn gitignore_honored_without_git_dir() {
-        // No `.git` anywhere — `.gitignore` is still applied (require_git=false).
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::create_dir(proj.join("gen")).unwrap();
-        std::fs::write(proj.join(".gitignore"), "gen/\n").unwrap();
-        std::fs::write(proj.join("gen/derived.py"), "def d():\n    pass\n").unwrap();
-        std::fs::write(proj.join("src.py"), "def s():\n    pass\n").unwrap();
-
-        index_dir_filtered(&db, &proj, &WalkFilter::unrestricted()).unwrap();
-
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.ends_with("src.py")));
-        assert!(!files.iter().any(|f| f.contains("gen")));
-    }
-
-    #[test]
-    fn respect_gitignore_false_indexes_ignored_but_keeps_floor() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::create_dir(proj.join("gen")).unwrap();
-        std::fs::write(proj.join(".gitignore"), "gen/\n").unwrap();
-        std::fs::write(proj.join("gen/derived.py"), "def d():\n    pass\n").unwrap();
-        std::fs::create_dir(proj.join("node_modules")).unwrap();
-        std::fs::write(proj.join("node_modules/x.py"), "def x():\n    pass\n").unwrap();
-
-        let filter = WalkFilter {
-            exclude: ExcludeGlobs::empty(),
-            respect_gitignore: false,
-            ..Default::default()
-        };
-        index_dir_filtered(&db, &proj, &filter).unwrap();
-
-        let files = db.all_files().unwrap();
-        // Opt-out indexes the gitignored file...
-        assert!(
-            files.iter().any(|f| f.contains("gen")),
-            "respect_gitignore=false should index gen/; got {files:?}"
-        );
-        // ...but the floor still skips node_modules.
-        assert!(!files.iter().any(|f| f.contains("node_modules")));
-    }
-
-    #[test]
-    fn cartogignore_is_honored() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        std::fs::create_dir(proj.join("secret")).unwrap();
-        // .cartogignore uses gitignore syntax and is always honored.
-        std::fs::write(proj.join(".cartogignore"), "secret/\n").unwrap();
-        std::fs::write(proj.join("secret/s.py"), "def s():\n    pass\n").unwrap();
-        std::fs::write(proj.join("ok.py"), "def ok():\n    pass\n").unwrap();
-
-        index_dir_filtered(&db, &proj, &WalkFilter::unrestricted()).unwrap();
-
-        let files = db.all_files().unwrap();
-        assert!(files.iter().any(|f| f.ends_with("ok.py")));
-        assert!(!files.iter().any(|f| f.contains("secret")));
-    }
-
-    #[test]
-    fn gitignore_and_exclude_compose() {
-        let db = cartog_db::Database::open_memory().unwrap();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let proj = visible_dir(&tmp, "proj");
-        for d in ["a", "b", "c"] {
-            std::fs::create_dir(proj.join(d)).unwrap();
-            std::fs::write(proj.join(d).join("f.py"), "def f():\n    pass\n").unwrap();
-        }
-        std::fs::write(proj.join(".gitignore"), "a/\n").unwrap(); // a/ via gitignore
-
-        let filter = WalkFilter {
-            exclude: ExcludeGlobs::from_globs(&["b/**".to_string()]).unwrap(), // b/ via exclude
-            respect_gitignore: true,
-            ..Default::default()
-        };
-        index_dir_filtered(&db, &proj, &filter).unwrap();
-
-        // Stored paths are repo-root-relative (e.g. "c/f.py").
-        let files = db.all_files().unwrap();
-        assert!(!files.iter().any(|f| f.starts_with("a/")), "a/ gitignored");
-        assert!(!files.iter().any(|f| f.starts_with("b/")), "b/ excluded");
-        assert!(files.iter().any(|f| f.starts_with("c/")), "c/ kept");
-    }
-
-    #[test]
-    fn test_file_hash_deterministic() {
-        let h1 = file_hash("def foo(): pass");
-        let h2 = file_hash("def foo(): pass");
-        assert_eq!(h1, h2);
-    }
-
-    #[test]
-    fn test_file_hash_different_content() {
-        let h1 = file_hash("def foo(): pass");
-        let h2 = file_hash("def bar(): pass");
-        assert_ne!(h1, h2);
-    }
-
-    #[test]
-    fn extract_with_cached_returns_none_for_unregistered_language() {
-        assert!(extract_with_cached("klingon", "irrelevant", "a.kl").is_none());
-    }
-
-    #[test]
-    fn index_summary_reports_file_symbol_and_edge_counts() {
-        let r = IndexResult {
-            files_indexed: 3,
-            files_skipped: 1,
-            symbols_added: 12,
-            edges_added: 20,
-            edges_resolved: 18,
-            ..Default::default()
-        };
-        let s = render_index_summary(&r);
-        assert!(s.contains("Indexed 3 files (1 skipped, 0 removed)"));
-        assert!(s.contains("12 symbols"));
-        assert!(s.contains("20 edges (18 resolved)"));
-    }
-
-    #[test]
-    fn index_summary_shows_detail_for_removal_only_delta() {
-        // A pass that only removes symbols (no new/modified/unchanged) must
-        // still report the removed count, not a bare "0 symbols".
-        let r = IndexResult {
-            files_indexed: 1,
-            symbols_removed: 4,
-            ..Default::default()
-        };
-        let s = render_index_summary(&r);
-        assert!(
-            s.contains("4 removed"),
-            "removal-only delta must surface the removed count: {s}"
-        );
-    }
-
-    #[test]
-    fn index_summary_breaks_out_lsp_resolution_when_present() {
-        let r = IndexResult {
-            files_indexed: 1,
-            symbols_added: 5,
-            edges_added: 10,
-            edges_resolved: 6,
-            edges_lsp_resolved: 3,
-            edges_marked_external: 1,
-            ..Default::default()
-        };
-        let s = render_index_summary(&r);
-        assert!(s.contains("9 resolved"), "6 heuristic + 3 LSP = 9");
-        assert!(s.contains("6 heuristic + 3 LSP"));
-        assert!(s.contains("1 external"));
-    }
-
-    #[test]
-    fn index_summary_lists_unsupported_languages() {
-        let r = IndexResult {
-            files_indexed: 2,
-            files_unsupported: 4,
-            unsupported_by_ext: vec![("kt".into(), 3), ("cpp".into(), 1)],
-            ..Default::default()
-        };
-        let s = render_index_summary(&r);
-        assert!(s.contains("4 files in unsupported languages"));
-        assert!(s.contains("3 .kt"));
-    }
-
-    #[test]
-    fn extract_with_cached_extracts_for_known_language() {
-        let result = extract_with_cached("python", "def foo():\n    pass\n", "a.py")
-            .expect("python is registered")
-            .expect("valid source extracts");
-        assert!(
-            result.symbols.iter().any(|s| s.name == "foo"),
-            "expected `foo` among extracted symbols"
-        );
-    }
-
-    #[test]
-    fn test_is_ignored_directories() {
-        let ignored_dirs = [
-            ".git",
-            "node_modules",
-            "__pycache__",
-            "target",
-            "dist",
-            "build",
-            ".venv",
-            "var",
-            "builds",
-        ];
-        for name in ignored_dirs {
-            assert!(is_ignored(name, true, 1), "{name} should be ignored");
-        }
-        for name in ["src", "lib", "tests", "docs"] {
-            assert!(!is_ignored(name, true, 1), "{name} should NOT be ignored");
-        }
-        // Files are never ignored by the floor.
-        assert!(!is_ignored("node_modules", false, 1));
-    }
-
-    #[test]
-    fn test_var_and_builds_not_ignored_when_nested() {
-        // "var"/"builds" are floored only at depth 1 (project root); a nested
-        // `src/var` (depth 2) is valid application code.
-        assert!(is_ignored("var", true, 1));
-        assert!(is_ignored("builds", true, 1));
-        assert!(!is_ignored("var", true, 2), "src/var should NOT be ignored");
-        assert!(
-            !is_ignored("builds", true, 2),
-            "src/builds should NOT be ignored"
-        );
-    }
-
-    #[test]
-    fn test_git_changed_files_no_commit() {
-        // When last_commit is None, should return None (first index → full scan)
-        let result = git_changed_files(Path::new("."), None);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_git_changed_files_invalid_commit() {
-        // A commit hash that doesn't exist should return None (fallback to hash)
-        let result = git_changed_files(
-            Path::new("."),
-            Some("0000000000000000000000000000000000000000"),
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_git_changed_files_valid_head() {
-        // If we diff HEAD against HEAD, the changed set should be empty
-        // (only working tree / untracked files would appear)
-        let head = git_head_commit(Path::new("."));
-        if let Some(commit) = head {
-            let result = git_changed_files(Path::new("."), Some(&commit));
-            // Should return Some (valid commit), though the set may contain untracked/modified files
-            assert!(result.is_some());
-        }
-    }
-
-    #[test]
-    fn db_sidecars_are_recognized() {
-        assert!(is_db_sidecar(".cartog.db"));
-        assert!(is_db_sidecar(".cartog.db-wal"));
-        assert!(is_db_sidecar(".cartog.db-shm"));
-        assert!(is_db_sidecar("sub/db.sqlite"));
-        assert!(is_db_sidecar("db.sqlite-wal"));
-        assert!(!is_db_sidecar("main.rs"));
-        assert!(!is_db_sidecar("app.dart"));
-    }
-
-    #[test]
-    fn unsupported_files_are_counted_not_silently_dropped() {
-        use cartog_db::Database;
-        // TempDir names start with '.', which the walker treats as hidden and
-        // prunes — nest a non-dot project dir so the walk descends into it.
-        let tmp = tempfile::TempDir::new().unwrap();
-        let dir = tmp.path().join("proj");
-        std::fs::create_dir(&dir).unwrap();
-        std::fs::write(dir.join("a.rs"), "fn main() {}\n").unwrap();
-        std::fs::write(dir.join("b.cs"), "class P {}\n").unwrap();
-        std::fs::write(dir.join("c.cs"), "class Q {}\n").unwrap();
-        std::fs::write(dir.join("d.cpp"), "int main() {}\n").unwrap();
-        // cartog's own DB sidecars must NOT count as unsupported languages.
-        std::fs::write(dir.join(".cartog.db"), "x").unwrap();
-        std::fs::write(dir.join(".cartog.db-wal"), "x").unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let r = index_directory(
-            &db,
-            &dir,
-            true,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        assert_eq!(r.files_indexed, 1, "only a.rs is supported");
-        assert_eq!(
-            r.files_unsupported, 3,
-            "2 csharp + 1 cpp, db sidecars excluded"
-        );
-        // Descending by count, ties broken alphabetically.
-        assert_eq!(
-            r.unsupported_by_ext,
-            vec![("cs".to_string(), 2), ("cpp".to_string(), 1)]
-        );
-    }
-
-    #[test]
-    fn dotted_import_leaf_is_separator_escaped_in_stored_id() {
-        // Real-output guard for the symbol-ID escaping fix: a dotted import leaf
-        // (os.path) is stored with its separator escaped (os%2Epath), not raw.
-        // The parented method (os.path) keeps its raw structural id; the two
-        // differ in the kind segment, so this checks escaping is applied in
-        // stored output — not a same-kind collision (that's the cartog-core
-        // injectivity unit test).
-        use cartog_db::Database;
-
-        // TempDir names start with '.', which the walker prunes as hidden — nest
-        // a non-dot project dir so the walk descends into it.
-        let tmp = tempfile::TempDir::new().unwrap();
-        let dir = tmp.path().join("proj");
-        std::fs::create_dir(&dir).unwrap();
-        std::fs::write(
-            dir.join("a.py"),
-            "import os.path\n\nclass os:\n    def path(self):\n        return 1\n",
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &dir,
-            true,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let ids: Vec<String> = db
-            .outline("a.py")
-            .unwrap()
-            .into_iter()
-            .map(|s| s.id)
-            .collect();
-
-        assert!(
-            ids.contains(&"a.py:import:os%2Epath".to_string()),
-            "dotted import leaf must be escaped: {ids:?}"
-        );
-        assert!(
-            ids.contains(&"a.py:method:os.path".to_string()),
-            "parented method keeps its raw structural id: {ids:?}"
-        );
-        // The escaped import id is stored verbatim — the raw dotted form must not
-        // appear (that would mean escaping was skipped on the stored path).
-        assert!(
-            !ids.contains(&"a.py:import:os.path".to_string()),
-            "raw (unescaped) dotted import id must not be stored: {ids:?}"
-        );
-    }
-
-    #[test]
-    fn test_index_directory_force() {
-        use cartog_db::Database;
-
-        let db = Database::open_memory().unwrap();
-        let fixtures = Path::new(env!("CARGO_MANIFEST_DIR")).join("../../tests/fixtures/auth");
-
-        if fixtures.exists() {
-            // First index
-            let r1 = index_directory(
-                &db,
-                &fixtures,
-                false,
-                false,
-                None,
-                None,
-                crate::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &crate::WalkFilter::unrestricted(),
-            )
-            .unwrap();
-            assert!(r1.files_indexed > 0);
-            assert!(r1.dirty_files > 0);
-
-            // Second index without force — should skip all files (no-op)
-            let r2 = index_directory(
-                &db,
-                &fixtures,
-                false,
-                false,
-                None,
-                None,
-                crate::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &crate::WalkFilter::unrestricted(),
-            )
-            .unwrap();
-            assert_eq!(r2.files_indexed, 0);
-            assert!(r2.files_skipped > 0);
-            assert_eq!(
-                r2.dirty_files, 0,
-                "no-op reindex must report zero dirty files — gates the LSP pass"
-            );
-            assert_eq!(
-                r2.edges_lsp_resolved, 0,
-                "no-op reindex must not run LSP resolution"
-            );
-
-            // Force re-index — dirty_files matches files_indexed
-            let r3 = index_directory(
-                &db,
-                &fixtures,
-                true,
-                false,
-                None,
-                None,
-                crate::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &crate::WalkFilter::unrestricted(),
-            )
-            .unwrap();
-            assert_eq!(r3.files_indexed, r1.files_indexed);
-            assert_eq!(r3.files_skipped, 0);
-            assert_eq!(r3.dirty_files, r3.files_indexed);
-        }
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn test_noop_reindex_does_not_run_lsp() {
-        // Regression guard: no dirty files → no LSP pass.
-        use cartog_db::Database;
-
-        let db = Database::open_memory().unwrap();
-        let fixtures = Path::new(env!("CARGO_MANIFEST_DIR")).join("../../tests/fixtures/auth");
-        if !fixtures.exists() {
-            return;
-        }
-
-        // Prime, then re-run. Don't assert LSP found anything (depends on
-        // whether pyright is on PATH in CI) — only that the second pass skips it.
-        let _ = index_directory(
-            &db,
-            &fixtures,
-            false,
-            true,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let r2 = index_directory(
-            &db,
-            &fixtures,
-            false,
-            true,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert_eq!(r2.dirty_files, 0);
-        assert_eq!(r2.edges_lsp_resolved, 0);
-    }
-
-    #[test]
-    fn no_lsp_index_seals_unresolved_edges_at_state_four() {
-        // Watch / --no-lsp runs (lsp=false) have no LSP pass to retry the
-        // heuristic's leftovers, so a state=0 edge would be re-walked on every
-        // re-index (#109 amplification). The indexer must seal it at state=4.
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(root.join("a.py"), "def caller():\n    nowhere('x')\n").unwrap();
-
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            false,
-            false, // lsp off — the only resolver is the heuristic
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // The unresolvable `nowhere` call is sealed, so it no longer surfaces
-        // to the unresolved set the next pass would scan.
-        let names: Vec<String> = db
-            .unresolved_edges()
-            .unwrap()
-            .into_iter()
-            .map(|e| e.target_name)
-            .collect();
-        assert!(
-            !names.contains(&"nowhere".to_string()),
-            "no-lsp index must seal the unresolvable edge at state=4, got {names:?}"
-        );
-    }
-
-    #[test]
-    fn test_added_symbol_reopens_unresolvable_edges() {
-        // Name-keyed reset: a new symbol whose name matches a state=2 edge
-        // returns the edge to state=0 (or state=1 if the heuristic resolves it).
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        // Rust's tempfile creates `.tmpXXXX` directories on macOS — the leading
-        // dot makes is_ignored() reject the walk root. Nest a non-dotted child.
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(root.join("a.py"), "def caller():\n    find_user('x')\n").unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let r1 = index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert!(
-            r1.files_indexed >= 1,
-            "expected a.py to index, got {:?}",
-            r1
-        );
-
-        // The no-lsp index seals leftovers at state=4; reopen so the rest of
-        // the test exercises the state=0 → marker lifecycle as before.
-        db.reopen_heuristic_exhausted().unwrap();
-        let before = db.unresolved_edges().unwrap();
-        let find_user = before
-            .iter()
-            .find(|e| e.target_name == "find_user")
-            .expect("find_user edge should exist as unresolved after first index");
-        let edge_id = find_user.edge_id;
-        db.mark_edge_unresolvable(edge_id).unwrap();
-        assert!(db.is_edge_unresolvable(edge_id).unwrap());
-
-        // Adding b.py with find_user definition should reopen the marker.
-        std::fs::write(root.join("b.py"), "def find_user(name):\n    return None\n").unwrap();
-        index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        assert!(
-            !db.is_edge_unresolvable(edge_id).unwrap(),
-            "edge must not stay state=2 after a matching target appears"
-        );
-    }
-
-    #[test]
-    fn reindex_invalidates_embedding_of_modified_symbol() {
-        // Drift regression: a symbol whose body changes keeps its stable id, so
-        // its old embedding must be dropped on re-index — otherwise
-        // symbols_needing_embeddings() skips it and the vector stays stale.
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        // Body must exceed MIN_CONTENT_BYTES (50) so content is extracted.
-        std::fs::write(
-            root.join("a.py"),
-            "def greet(name):\n    message = 'hello there ' + name\n    return message\n",
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let idx = |db: &Database| {
-            index_directory(
-                db,
-                &root,
-                false,
-                false,
-                None,
-                None,
-                crate::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &crate::WalkFilter::unrestricted(),
-            )
-            .unwrap()
-        };
-        idx(&db);
-
-        // Simulate a prior `rag index`: embed the only content symbol.
-        let needing = db.symbols_needing_embeddings().unwrap();
-        assert_eq!(needing.len(), 1, "expected greet() to need embedding");
-        let greet_id = needing[0].clone();
-        let bytes: Vec<u8> = vec![0.0f32; 384]
-            .iter()
-            .flat_map(|f| f.to_le_bytes())
-            .collect();
-        let eid = db.get_or_create_embedding_id(&greet_id).unwrap();
-        db.upsert_embedding(eid, &bytes).unwrap();
-        assert!(db.symbols_needing_embeddings().unwrap().is_empty());
-
-        // Edit the body (same name/kind/file → same stable id) and re-index.
-        std::fs::write(
-            root.join("a.py"),
-            "def greet(name):\n    message = 'goodbye and farewell ' + name\n    return message\n",
-        )
-        .unwrap();
-        idx(&db);
-
-        // The drifted embedding is gone and the symbol re-enters the queue.
-        assert!(
-            !db.has_embedding(&greet_id).unwrap(),
-            "modified symbol's stale embedding must be cleared"
-        );
-        assert_eq!(
-            db.symbols_needing_embeddings().unwrap(),
-            vec![greet_id],
-            "modified symbol must re-enter the needs-embedding set"
-        );
-    }
-
-    #[test]
-    fn reindex_keeps_embedding_of_unchanged_sibling() {
-        // The drift-clear is scoped to modified symbols: a sibling whose own
-        // content is untouched keeps its embedding even when the file is dirty.
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        let greet_v1 =
-            "def greet(name):\n    message = 'hello there ' + name\n    return message\n";
-        let farewell = "def farewell(name):\n    message = 'goodbye and take care ' + name\n    return message\n";
-        std::fs::write(root.join("a.py"), format!("{greet_v1}\n\n{farewell}")).unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let idx = |db: &Database| {
-            index_directory(
-                db,
-                &root,
-                false,
-                false,
-                None,
-                None,
-                crate::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &crate::WalkFilter::unrestricted(),
-            )
-            .unwrap()
-        };
-        idx(&db);
-
-        let bytes: Vec<u8> = vec![0.0f32; 384]
-            .iter()
-            .flat_map(|f| f.to_le_bytes())
-            .collect();
-        for id in db.symbols_needing_embeddings().unwrap() {
-            let eid = db.get_or_create_embedding_id(&id).unwrap();
-            db.upsert_embedding(eid, &bytes).unwrap();
-        }
-        let farewell_id = db
-            .all_content_symbol_ids()
-            .unwrap()
-            .into_iter()
-            .find(|id| id.contains("farewell"))
-            .expect("farewell symbol id");
-
-        // Change only greet(); farewell()'s own content is identical.
-        let greet_v2 =
-            "def greet(name):\n    message = 'goodbye and farewell ' + name\n    return message\n";
-        std::fs::write(root.join("a.py"), format!("{greet_v2}\n\n{farewell}")).unwrap();
-        idx(&db);
-
-        assert!(
-            db.has_embedding(&farewell_id).unwrap(),
-            "unchanged sibling must keep its embedding"
-        );
-    }
-
-    #[test]
-    fn reindex_drops_stale_content_when_modified_body_shrinks_below_threshold() {
-        // A modified symbol whose new body no longer yields content must not keep
-        // its pre-edit content row (else it re-embeds stale text forever).
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        // v1 body is well over MIN_CONTENT_BYTES (50) so content is stored.
-        std::fs::write(
-            root.join("a.py"),
-            "def greet(name):\n    message = 'a long enough greeting for ' + name\n    return message\n",
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let idx = |db: &Database| {
-            index_directory(
-                db,
-                &root,
-                false,
-                false,
-                None,
-                None,
-                crate::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &crate::WalkFilter::unrestricted(),
-            )
-            .unwrap()
-        };
-        idx(&db);
-        let greet_id = db.symbols_needing_embeddings().unwrap()[0].clone();
-        assert!(db.get_symbol_content(&greet_id).unwrap().is_some());
-
-        // Shrink the body below the content threshold; same stable id (same name).
-        std::fs::write(root.join("a.py"), "def greet(name):\n    pass\n").unwrap();
-        idx(&db);
-
-        assert!(
-            db.get_symbol_content(&greet_id).unwrap().is_none(),
-            "stale content row must be deleted when the modified body loses content"
-        );
-        assert!(
-            !db.symbols_needing_embeddings().unwrap().contains(&greet_id),
-            "a symbol with no content must not be queued for embedding"
-        );
-    }
-
-    #[test]
-    fn test_force_reindex_does_not_inherit_sticky_markers() {
-        // End-to-end contract: --force is the documented escape hatch for
-        // "retry everything". Under --force, every file is re-parsed and
-        // `clear_edges_for_file` / `clear_file_data_in_tx` wipe the edge
-        // rows before `resolve_edges` runs — so the post-force edges have
-        // fresh auto-increment IDs at default state=0 regardless of what
-        // state the pre-force edges held. This test exercises that path
-        // through real indexing; the targeted unit test for the SQL filter
-        // (`IN (2, 3)`) lives in cartog-db
-        // (test_reset_all_unresolvable_resets_state_two_and_three).
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(
-            root.join("a.py"),
-            "def caller():\n    find_x()\n    find_ext()\n",
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // The no-lsp index seals leftovers at state=4; reopen so we can mark
-        // them {2, 3} and test that --force clears those sticky markers.
-        db.reopen_heuristic_exhausted().unwrap();
-        let pre = db.unresolved_edges().unwrap();
-        let find_x_id = pre
-            .iter()
-            .find(|e| e.target_name == "find_x")
-            .expect("find_x edge should exist")
-            .edge_id;
-        let find_ext_id = pre
-            .iter()
-            .find(|e| e.target_name == "find_ext")
-            .expect("find_ext edge should exist")
-            .edge_id;
-        db.mark_edge_unresolvable(find_x_id).unwrap();
-        db.mark_edge_external(find_ext_id).unwrap();
-
-        // --force = true: rebuilds edges with fresh IDs, must NOT inherit state {2, 3}.
-        index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        // --force clears the sticky LSP markers (reset_all_unresolvable), then
-        // this no-lsp run re-seals the still-unresolvable edges at state=4 — a
-        // fresh heuristic-exhaustion verdict, not the inherited {2, 3} markers.
-        // Reopen state=4 to confirm both edges came back to the resolvable set.
-        db.reopen_heuristic_exhausted().unwrap();
-        let post = db.unresolved_edges().unwrap();
-        assert!(
-            post.iter().any(|e| e.target_name == "find_x"),
-            "after --force, find_x must not stay at an inherited {{2, 3}} marker"
-        );
-        assert!(
-            post.iter().any(|e| e.target_name == "find_ext"),
-            "after --force, find_ext must not stay at an inherited {{2, 3}} marker"
-        );
-    }
-
-    #[test]
-    fn test_name_keyed_reset_reopens_external_edges() {
-        // If an edge was marked state=3 (target outside the indexed root) and
-        // the user then vendors that target in-tree, indexing the new file
-        // must reopen the external marker so LSP retries it.
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(root.join("a.py"), "def caller():\n    vendored_helper()\n").unwrap();
-
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // The no-lsp index seals leftovers at state=4; reopen so we can mark
-        // the edge external and test the vendored-in-tree reopen path.
-        db.reopen_heuristic_exhausted().unwrap();
-        let unresolved = db.unresolved_edges().unwrap();
-        let edge_id = unresolved
-            .iter()
-            .find(|e| e.target_name == "vendored_helper")
-            .expect("vendored_helper edge should exist")
-            .edge_id;
-        db.mark_edge_external(edge_id).unwrap();
-        assert_eq!(db.edge_resolution_state(edge_id).unwrap(), 3);
-
-        // Vendor the dep in-tree.
-        std::fs::write(
-            root.join("vendor.py"),
-            "def vendored_helper():\n    return 1\n",
-        )
-        .unwrap();
-        index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // After the name-keyed reset, the edge is reopened (state=0) and the
-        // heuristic resolver runs in the same indexing pass — `vendored_helper`
-        // is now defined in the same directory, so the same-dir heuristic
-        // resolves the edge to state=1. Asserting state=1 (not just "not 3")
-        // catches a future regression that breaks the reset-then-resolve
-        // pipeline (e.g. silently re-marking as state=2).
-        assert_eq!(
-            db.edge_resolution_state(edge_id).unwrap(),
-            1,
-            "vendored target should reopen the external marker AND be resolved by the heuristic"
-        );
-    }
-
-    #[test]
-    fn test_noop_reindex_preserves_unresolvable_and_external_markers() {
-        // Defensive: a no-op reindex must not touch state {2, 3} markers — no
-        // spurious resets (would burn the gate), no spurious re-marks.
-        use cartog_db::Database;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(
-            root.join("a.py"),
-            "def caller():\n    find_x()\n    find_y()\n",
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // The no-lsp index seals leftovers at state=4; reopen so we can mark
-        // them {2, 3} and test that a no-op reindex preserves those markers.
-        db.reopen_heuristic_exhausted().unwrap();
-        let unresolved = db.unresolved_edges().unwrap();
-        let burned = unresolved
-            .iter()
-            .find(|e| e.target_name == "find_x")
-            .expect("find_x edge should exist");
-        let ext = unresolved
-            .iter()
-            .find(|e| e.target_name == "find_y")
-            .expect("find_y edge should exist");
-        db.mark_edge_unresolvable(burned.edge_id).unwrap();
-        db.mark_edge_external(ext.edge_id).unwrap();
-
-        // No file changes → reindex is a no-op → markers survive.
-        let r = index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert_eq!(r.dirty_files, 0);
-
-        assert_eq!(
-            db.edge_resolution_state(burned.edge_id).unwrap(),
-            2,
-            "no-op reindex must not reset state=2"
-        );
-        assert_eq!(
-            db.edge_resolution_state(ext.edge_id).unwrap(),
-            3,
-            "no-op reindex must not reset state=3"
-        );
-    }
-
-    // ── Dedup tests ──
-
-    #[test]
-    fn test_dedup_3way_collision_preserves_invariant() {
-        // Three symbols with the same stable id — simulates conditional
-        // redefinitions (e.g. `if/elif/else: def foo`).
-        let mk_sym = || {
-            cartog_core::Symbol::new(
-                "foo",
-                cartog_core::SymbolKind::Function,
-                "test.py",
-                1,
-                2,
-                0,
-                10,
-                None,
-            )
-        };
-        let base_id = mk_sym().id.clone();
-        let mut symbols = vec![mk_sym(), mk_sym(), mk_sym()];
-        let mut edges = vec![
-            cartog_core::Edge::new(
-                base_id.clone(),
-                "bar",
-                cartog_core::EdgeKind::Calls,
-                "test.py",
-                1,
-            ),
-            cartog_core::Edge::new(
-                base_id.clone(),
-                "baz",
-                cartog_core::EdgeKind::Calls,
-                "test.py",
-                2,
-            ),
-        ];
-
-        dedup_symbol_ids(&mut symbols, &mut edges);
-
-        // All three ids must now be distinct.
-        let ids: std::collections::HashSet<_> = symbols.iter().map(|s| s.id.as_str()).collect();
-        assert_eq!(ids.len(), 3, "3-way collision should produce 3 unique ids");
-
-        // First instance keeps the short id; 2nd and 3rd get numeric suffixes.
-        assert_eq!(symbols[0].id, base_id);
-        assert_eq!(symbols[1].id, format!("{base_id}:2"));
-        assert_eq!(symbols[2].id, format!("{base_id}:3"));
-
-        // Invariant: every edge.source_id must resolve to a surviving symbol.
-        for edge in &edges {
-            assert!(
-                ids.contains(edge.source_id.as_str()),
-                "edge source_id {:?} has no matching symbol after dedup",
-                edge.source_id
-            );
-        }
-    }
-
-    #[test]
-    fn test_dedup_no_collision_leaves_ids_unchanged() {
-        let mut symbols = vec![
-            cartog_core::Symbol::new(
-                "a",
-                cartog_core::SymbolKind::Function,
-                "f.py",
-                1,
-                2,
-                0,
-                10,
-                None,
-            ),
-            cartog_core::Symbol::new(
-                "b",
-                cartog_core::SymbolKind::Function,
-                "f.py",
-                3,
-                4,
-                11,
-                20,
-                None,
-            ),
-        ];
-        let id_a = symbols[0].id.clone();
-        let id_b = symbols[1].id.clone();
-        let mut edges: Vec<cartog_core::Edge> = vec![];
-        dedup_symbol_ids(&mut symbols, &mut edges);
-        assert_eq!(symbols[0].id, id_a);
-        assert_eq!(symbols[1].id, id_b);
-    }
-
-    // ── Merkle hashing tests ──
-
-    #[test]
-    fn test_compute_merkle_hashes_populates_fields() {
-        let source = "def foo():\n    pass\n";
-        let mut symbols = vec![cartog_core::Symbol::new(
-            "foo",
-            cartog_core::SymbolKind::Function,
-            "test.py",
-            1,
-            2,
-            0,
-            source.len() as u32,
-            None,
-        )];
-
-        compute_merkle_hashes(&mut symbols, source);
-
-        assert!(symbols[0].content_hash.is_some());
-        assert!(symbols[0].subtree_hash.is_some());
-    }
-
-    #[test]
-    fn test_merkle_hashes_stable_across_position_changes() {
-        let source_v1 = "def foo():\n    pass\n";
-        let source_v2 = "\n\ndef foo():\n    pass\n";
-
-        let mut sym_v1 = vec![cartog_core::Symbol::new(
-            "foo",
-            cartog_core::SymbolKind::Function,
-            "test.py",
-            1,
-            2,
-            0,
-            source_v1.len() as u32,
-            None,
-        )];
-        let mut sym_v2 = vec![cartog_core::Symbol::new(
-            "foo",
-            cartog_core::SymbolKind::Function,
-            "test.py",
-            3,
-            4,
-            2,
-            source_v2.len() as u32,
-            None,
-        )];
-
-        compute_merkle_hashes(&mut sym_v1, source_v1);
-        compute_merkle_hashes(&mut sym_v2, source_v2);
-
-        // content_hash depends on body text — different offset means different body slice
-        // but if the body text is the same, hashes should match
-        // Here the body text is the same "def foo():\n    pass\n"
-        assert_eq!(sym_v1[0].content_hash, sym_v2[0].content_hash);
-    }
-
-    #[test]
-    fn test_merkle_diff_detects_added_symbol() {
-        let old_hashes: Vec<(String, Option<String>, Option<String>)> = vec![];
-
-        let mut new_symbols = vec![cartog_core::Symbol::new(
-            "foo",
-            cartog_core::SymbolKind::Function,
-            "test.py",
-            1,
-            5,
-            0,
-            50,
-            None,
-        )];
-        new_symbols[0].content_hash = Some("abc".to_string());
-        new_symbols[0].subtree_hash = Some("def".to_string());
-
-        let diff = merkle_diff(&new_symbols, &old_hashes);
-        assert_eq!(diff.added.len(), 1);
-        assert_eq!(diff.removed.len(), 0);
-        assert_eq!(diff.modified.len(), 0);
-    }
-
-    #[test]
-    fn test_merkle_diff_detects_removed_symbol() {
-        let old_hashes = vec![(
-            "test.py:function:foo".to_string(),
-            Some("abc".to_string()),
-            Some("def".to_string()),
-        )];
-
-        let new_symbols: Vec<cartog_core::Symbol> = vec![];
-
-        let diff = merkle_diff(&new_symbols, &old_hashes);
-        assert_eq!(diff.added.len(), 0);
-        assert_eq!(diff.removed.len(), 1);
-        assert_eq!(diff.removed[0], "test.py:function:foo");
-    }
-
-    #[test]
-    fn test_merkle_diff_detects_unchanged() {
-        let old_hashes = vec![(
-            "test.py:function:foo".to_string(),
-            Some("abc".to_string()),
-            Some("def".to_string()),
-        )];
-
-        let mut new_symbols = vec![cartog_core::Symbol::new(
-            "foo",
-            cartog_core::SymbolKind::Function,
-            "test.py",
-            1,
-            5,
-            0,
-            50,
-            None,
-        )];
-        new_symbols[0].content_hash = Some("abc".to_string());
-        new_symbols[0].subtree_hash = Some("def".to_string());
-
-        let diff = merkle_diff(&new_symbols, &old_hashes);
-        assert_eq!(diff.unchanged, 1);
-        assert_eq!(diff.added.len(), 0);
-        assert_eq!(diff.modified.len(), 0);
-    }
-
-    #[test]
-    fn test_merkle_diff_detects_modified() {
-        let old_hashes = vec![(
-            "test.py:function:foo".to_string(),
-            Some("old_hash".to_string()),
-            Some("old_subtree".to_string()),
-        )];
-
-        let mut new_symbols = vec![cartog_core::Symbol::new(
-            "foo",
-            cartog_core::SymbolKind::Function,
-            "test.py",
-            1,
-            5,
-            0,
-            50,
-            None,
-        )];
-        new_symbols[0].content_hash = Some("new_hash".to_string());
-        new_symbols[0].subtree_hash = Some("new_subtree".to_string());
-
-        let diff = merkle_diff(&new_symbols, &old_hashes);
-        assert_eq!(diff.modified.len(), 1);
-        assert_eq!(diff.unchanged, 0);
-    }
-
-    // ── Integration test: full incremental pipeline ──
-
-    #[test]
-    fn test_incremental_merkle_diff_pipeline() {
-        use cartog_db::Database;
-
-        let tmp = tempfile::TempDir::new().unwrap();
-        // Create a non-dot subdirectory (tempfile may create .tmpXXX on macOS,
-        // which is_ignored_dirname skips)
-        let dir = tmp.path().join("project");
-        std::fs::create_dir(&dir).unwrap();
-
-        // Initial files
-        let a_py = dir.join("a.py");
-        let b_py = dir.join("b.py");
-
-        std::fs::write(
-            &a_py,
-            r#"class Greeter:
-    def hello(self):
-        return "hi"
-    def goodbye(self):
-        return "bye"
-"#,
-        )
-        .unwrap();
-
-        std::fs::write(
-            &b_py,
-            r#"from a import Greeter
-def main():
-    g = Greeter()
-    g.hello()
-"#,
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-
-        // ── Index 1: initial full index ──
-        let r1 = index_directory(
-            &db,
-            &dir,
-            true,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert_eq!(r1.files_indexed, 2);
-        assert!(r1.symbols_added > 0, "should have symbols");
-
-        let outline_a = db.outline("a.py").unwrap();
-        assert_eq!(outline_a.len(), 3, "Greeter + hello + goodbye");
-        let names_a: Vec<&str> = outline_a.iter().map(|s| s.name.as_str()).collect();
-        assert!(names_a.contains(&"Greeter"));
-        assert!(names_a.contains(&"hello"));
-        assert!(names_a.contains(&"goodbye"));
-
-        // Capture stable IDs
-        let hello_id_v1 = outline_a
-            .iter()
-            .find(|s| s.name == "hello")
-            .unwrap()
-            .id
-            .clone();
-        let greeter_id_v1 = outline_a
-            .iter()
-            .find(|s| s.name == "Greeter")
-            .unwrap()
-            .id
-            .clone();
-
-        // Verify Merkle hashes populated
-        let hashes = db.get_symbol_hashes_for_file("a.py").unwrap();
-        assert!(
-            hashes
-                .iter()
-                .all(|(_, ch, sh)| ch.is_some() && sh.is_some()),
-            "all symbols should have hashes after indexing"
-        );
-
-        // ── Index 2: add a function to a.py ──
-        std::fs::write(
-            &a_py,
-            r#"class Greeter:
-    def hello(self):
-        return "hi"
-    def goodbye(self):
-        return "bye"
-
-def standalone():
-    return "I am new"
-"#,
-        )
-        .unwrap();
-
-        let r2 = index_directory(
-            &db,
-            &dir,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert_eq!(r2.files_indexed, 1, "only a.py changed");
-        assert!(r2.files_skipped > 0, "b.py should be skipped");
-        assert_eq!(r2.symbols_added, 1, "standalone is new");
-        assert!(
-            r2.symbols_unchanged >= 2,
-            "hello and goodbye should be unchanged, got {}",
-            r2.symbols_unchanged
-        );
-
-        let outline_a2 = db.outline("a.py").unwrap();
-        assert_eq!(
-            outline_a2.len(),
-            4,
-            "Greeter + hello + goodbye + standalone"
-        );
-        assert!(outline_a2.iter().any(|s| s.name == "standalone"));
-
-        // Verify ID stability: hello and Greeter keep same IDs
-        let hello_id_v2 = outline_a2
-            .iter()
-            .find(|s| s.name == "hello")
-            .unwrap()
-            .id
-            .clone();
-        let greeter_id_v2 = outline_a2
-            .iter()
-            .find(|s| s.name == "Greeter")
-            .unwrap()
-            .id
-            .clone();
-        assert_eq!(hello_id_v1, hello_id_v2, "hello ID should be stable");
-        assert_eq!(greeter_id_v1, greeter_id_v2, "Greeter ID should be stable");
-
-        // ── Index 3: remove goodbye from a.py ──
-        std::fs::write(
-            &a_py,
-            r#"class Greeter:
-    def hello(self):
-        return "hi"
-
-def standalone():
-    return "I am new"
-"#,
-        )
-        .unwrap();
-
-        let r3 = index_directory(
-            &db,
-            &dir,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert_eq!(r3.files_indexed, 1);
-        assert!(r3.symbols_removed >= 1, "goodbye should be removed");
-
-        let outline_a3 = db.outline("a.py").unwrap();
-        assert_eq!(outline_a3.len(), 3, "Greeter + hello + standalone");
-        assert!(
-            !outline_a3.iter().any(|s| s.name == "goodbye"),
-            "goodbye should be gone"
-        );
-
-        // hello ID still stable after removal of sibling
-        let hello_id_v3 = outline_a3
-            .iter()
-            .find(|s| s.name == "hello")
-            .unwrap()
-            .id
-            .clone();
-        assert_eq!(
-            hello_id_v1, hello_id_v3,
-            "hello ID stable after sibling removal"
-        );
-    }
-
-    // ── Integration test: Markdown document indexing ──
-
-    #[test]
-    fn test_markdown_indexing_end_to_end() {
-        use cartog_db::Database;
-
-        let tmp = tempfile::TempDir::new().unwrap();
-        let dir = tmp.path().join("project");
-        std::fs::create_dir(&dir).unwrap();
-
-        let md_file = dir.join("design.md");
-        std::fs::write(
-            &md_file,
-            r#"# Architecture
-
-This document describes the system architecture.
-
-## Authentication
-
-Users authenticate via JWT tokens. The server validates
-the token signature and checks expiration before granting access.
-
-## Database
-
-We use PostgreSQL with connection pooling via pgbouncer.
-"#,
-        )
-        .unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let result = index_directory(
-            &db,
-            &dir,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        assert_eq!(result.files_indexed, 1);
-        assert!(result.symbols_added >= 3, "should have at least 3 sections");
-
-        // Verify file entry
-        let file = db.get_file("design.md").unwrap();
-        assert!(file.is_some());
-        let file = file.unwrap();
-        assert_eq!(file.language, "markdown");
-
-        // Verify Document symbols exist
-        let outline = db.outline("design.md").unwrap();
-        assert!(
-            outline.len() >= 3,
-            "should have Architecture, Authentication, Database sections"
-        );
-
-        let names: Vec<&str> = outline.iter().map(|s| s.name.as_str()).collect();
-        assert!(
-            names.contains(&"Architecture"),
-            "missing Architecture section"
-        );
-        assert!(
-            names.contains(&"Authentication"),
-            "missing Authentication section"
-        );
-        assert!(names.contains(&"Database"), "missing Database section");
-
-        for sym in &outline {
-            assert_eq!(sym.kind, cartog_core::SymbolKind::Document);
-        }
-
-        // Verify symbol_content is populated
-        let auth_sym = outline.iter().find(|s| s.name == "Authentication").unwrap();
-        let content = db.get_symbol_content(&auth_sym.id).unwrap();
-        assert!(
-            content.is_some(),
-            "symbol_content should exist for document section"
-        );
-        let (text, header) = content.unwrap();
-        assert!(
-            text.contains("JWT tokens"),
-            "content should include section body"
-        );
-        assert!(
-            header.contains("Authentication"),
-            "header should include section name"
-        );
-    }
-
-    /// End-to-end Phase 3 atomicity test.
-    ///
-    /// Drives `index_directory` against a real fixture, then forces the DB to
-    /// run out of pages mid-Phase-3 by capping `max_page_count`. The expected
-    /// outcome: the index call returns `Err`, AND every Phase-3 write that
-    /// happened before the failure has been rolled back. The seed state is
-    /// established before the cap, so it must survive the rollback.
-    ///
-    /// This complements the cartog-db primitive tests (which prove that
-    /// `begin_indexing_tx` rolls back correctly): this test proves that
-    /// `index_directory` actually opens, uses, and rolls back the
-    /// transaction through every code path the real pipeline exercises
-    /// (per-file Merkle diff, deletion sweep, edge resolution, in-degree
-    /// compute, last_commit metadata write).
-    #[test]
-    fn test_index_directory_rolls_back_on_disk_full() {
-        use cartog_core::SymbolKind;
-        use cartog_db::Database;
-
-        let tmp = tempfile::TempDir::new().unwrap();
-        let dir = tmp.path().join("project");
-        std::fs::create_dir(&dir).unwrap();
-
-        // Seed file: get into a known indexed state under a generous page budget.
-        std::fs::write(dir.join("seed.py"), "def keep_me():\n    return 1\n").unwrap();
-
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &dir,
-            true,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .expect("seed index should succeed");
-
-        // Snapshot the seed state so we can assert it is preserved across the
-        // failed run.
-        let seed_outline = db.outline("seed.py").unwrap();
-        let seed_keep_me = seed_outline
-            .iter()
-            .find(|s| s.name == "keep_me")
-            .expect("seed symbol must be present after the first index");
-        let seed_keep_me_id = seed_keep_me.id.clone();
-
-        // Add a second file that the next `index_directory` call will try to
-        // ingest. Combined with a tight page cap, the new symbol/edge/content
-        // writes will hit `SQLITE_FULL` somewhere inside Phase 3.
-        std::fs::write(
-            dir.join("big.py"),
-            // Lots of small symbols: many independent INSERTs, so the page
-            // budget runs out partway through and the outer tx must roll back.
-            (0..200)
-                .map(|i| format!("def fn_{i}():\n    return {i}\n\n"))
-                .collect::<String>(),
-        )
-        .unwrap();
-
-        // Cap the DB at a page count that holds the seed comfortably but
-        // cannot fit the second file's worth of symbol/content rows.
-        // The exact value is empirical; on macOS APFS with the default 4 KiB
-        // page size, ~30 pages is enough to seed but not enough to ingest 200
-        // new functions through Phase 3.
-        db.set_max_page_count_for_tests(30).unwrap();
-
-        let result = index_directory(
-            &db,
-            &dir,
-            false,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        );
-        assert!(
-            result.is_err(),
-            "Phase 3 must fail when SQLite runs out of pages; got Ok({result:?})"
-        );
-
-        // Lift the cap so post-mortem queries can run.
-        db.set_max_page_count_for_tests(1_000_000).unwrap();
-
-        // Rollback assertions:
-        //
-        // 1. The seed symbol must still be there (a regression that wiped
-        //    pre-existing data is the worst flavor of the original bug).
-        // 2. None of the symbols from the failed file may have leaked through:
-        //    Phase 3 was wrapped in a single transaction, so partial writes
-        //    are rolled back atomically.
-        let seed_outline_after = db.outline("seed.py").unwrap();
-        assert!(
-            seed_outline_after.iter().any(|s| s.id == seed_keep_me_id),
-            "seed symbol must survive the rolled-back run"
-        );
-        let big_outline_after = db.outline("big.py").unwrap();
-        assert!(
-            big_outline_after.is_empty(),
-            "no symbols from the failed Phase 3 may persist; big.py outline: {:?}",
-            big_outline_after
-                .iter()
-                .map(|s| &s.name)
-                .collect::<Vec<_>>()
-        );
-
-        // The seed symbol is a function — a quick sanity check that the kind
-        // wasn't corrupted by partial writes either.
-        let kept = seed_outline_after
-            .iter()
-            .find(|s| s.id == seed_keep_me_id)
-            .unwrap();
-        assert_eq!(kept.kind, SymbolKind::Function);
-    }
-
-    // ── progress callback ──
-
-    fn tiny_python_project() -> (tempfile::TempDir, PathBuf) {
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(root.join("a.py"), "def f():\n    pass\n").unwrap();
-        std::fs::write(root.join("b.py"), "def g():\n    pass\n").unwrap();
-        (tmp, root)
-    }
-
-    #[test]
-    fn progress_callback_fires_in_phase_order() {
-        use cartog_db::Database;
-        use std::sync::Mutex;
-
-        let (_tmp, root) = tiny_python_project();
-        let db = Database::open_memory().unwrap();
-
-        let events: Mutex<Vec<ProgressUpdate>> = Mutex::new(Vec::new());
-        let cb = |u: ProgressUpdate| events.lock().unwrap().push(u);
-        let result = index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            Some(&cb),
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        assert!(result.files_indexed >= 2);
-        let events = events.into_inner().unwrap();
-        assert_eq!(events[0], ProgressUpdate::Walking);
-        // Each phase opens with done == 0 and later reports done == total.
-        assert!(events
-            .iter()
-            .any(|e| matches!(e, ProgressUpdate::Parsing { done: 0, total } if *total >= 2)));
-        assert!(events.iter().any(
-            |e| matches!(e, ProgressUpdate::Parsing { done, total } if done == total && *total >= 2)
-        ));
-        assert!(events
-            .iter()
-            .any(|e| matches!(e, ProgressUpdate::Storing { done: 0, total } if *total >= 2)));
-        assert!(events.iter().any(
-            |e| matches!(e, ProgressUpdate::Storing { done, total } if done == total && *total >= 2)
-        ));
-        // Phase order: last Walking < first Parsing < first Storing.
-        let pos = |pred: fn(&ProgressUpdate) -> bool| events.iter().position(pred).unwrap();
-        let walking = pos(|e| matches!(e, ProgressUpdate::Walking));
-        let parsing = pos(|e| matches!(e, ProgressUpdate::Parsing { .. }));
-        let storing = pos(|e| matches!(e, ProgressUpdate::Storing { .. }));
-        assert!(walking < parsing && parsing < storing);
-    }
-
-    #[test]
-    fn progress_callback_none_matches_some_for_result() {
-        use cartog_db::Database;
-
-        let (_t1, root1) = tiny_python_project();
-        let db1 = Database::open_memory().unwrap();
-        let r_none = index_directory(
-            &db1,
-            &root1,
-            true,
-            false,
-            None,
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let (_t2, root2) = tiny_python_project();
-        let db2 = Database::open_memory().unwrap();
-        let cb = |_: ProgressUpdate| {};
-        let r_some = index_directory(
-            &db2,
-            &root2,
-            true,
-            false,
-            Some(&cb),
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // Different temp dirs → different file modified-times can shift, but the
-        // count-based fields of IndexResult are deterministic on a fresh DB.
-        assert_eq!(r_none.files_indexed, r_some.files_indexed);
-        assert_eq!(r_none.symbols_added, r_some.symbols_added);
-        assert_eq!(r_none.edges_added, r_some.edges_added);
-    }
-
-    /// Progress callback emits Walking, then Parsing and Storing with positive
-    /// totals. Uses the in-repo fixture (the old env gate was set nowhere, so
-    /// the test was a silent no-op).
-    #[test]
-    fn progress_callback_emits_walking_then_parsing_and_storing() {
-        use cartog_db::Database;
-        use std::sync::Mutex;
-
-        let (_tmp, root) = tiny_python_project();
-        let db = Database::open_memory().unwrap();
-        let events: Mutex<Vec<ProgressUpdate>> = Mutex::new(Vec::new());
-        let cb = |u: ProgressUpdate| events.lock().unwrap().push(u);
-        index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            Some(&cb),
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let events = events.into_inner().unwrap();
-        assert!(
-            matches!(events.first(), Some(ProgressUpdate::Walking)),
-            "first progress event must be Walking, got {:?}",
-            events.first()
-        );
-        assert!(
-            events
-                .iter()
-                .any(|e| matches!(e, ProgressUpdate::Parsing { total, .. } if *total > 0)),
-            "must emit a Parsing event with a positive total"
-        );
-        assert!(
-            events
-                .iter()
-                .any(|e| matches!(e, ProgressUpdate::Storing { total, .. } if *total > 0)),
-            "must emit a Storing event with a positive total"
-        );
-    }
-
-    #[test]
-    fn progress_counter_climbs_mid_phase_for_large_repo() {
-        use cartog_db::Database;
-        use std::sync::Mutex;
-
-        // More than PROGRESS_STRIDE files so the in-loop stride emit fires at
-        // least one intermediate `done` (0 < done < total) — the path the small
-        // 2-file fixtures never exercise.
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        let n = PROGRESS_STRIDE * 3 + 5; // 197 files
-        for i in 0..n {
-            std::fs::write(
-                root.join(format!("m{i}.py")),
-                format!("def f{i}():\n    return {i}\n"),
-            )
-            .unwrap();
-        }
-
-        let db = Database::open_memory().unwrap();
-        let events: Mutex<Vec<ProgressUpdate>> = Mutex::new(Vec::new());
-        let cb = |u: ProgressUpdate| events.lock().unwrap().push(u);
-        index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            Some(&cb),
-            None,
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let events = events.into_inner().unwrap();
-        // A climbing parse event with done strictly between 0 and total.
-        assert!(
-            events.iter().any(|e| matches!(
-                e,
-                ProgressUpdate::Parsing { done, total } if *done > 0 && *done < *total
-            )),
-            "expected a mid-climb Parsing event, got {events:?}"
-        );
-        // Same for storing.
-        assert!(events.iter().any(|e| matches!(
-            e,
-            ProgressUpdate::Storing { done, total } if *done > 0 && *done < *total
-        )));
-        // Emitted parse `done` values never decrease (out-of-order rayon clamp).
-        let parse_dones: Vec<u32> = events
-            .iter()
-            .filter_map(|e| match e {
-                ProgressUpdate::Parsing { done, .. } => Some(*done),
-                _ => None,
-            })
-            .collect();
-        assert!(
-            parse_dones.windows(2).all(|w| w[0] <= w[1]),
-            "parse done must be non-decreasing, got {parse_dones:?}"
-        );
-    }
-
-    #[test]
-    fn cancel_probe_returning_true_aborts_with_cancelled_error() {
-        use cartog_db::Database;
-
-        let (_tmp, root) = tiny_python_project();
-        let db = Database::open_memory().unwrap();
-
-        let probe = || true;
-        let err = index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            Some(&probe),
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .expect_err("index must abort when probe trips at first phase boundary");
-        assert!(
-            err.to_string().contains("cancelled"),
-            "error must mention cancellation, got: {err}"
-        );
-    }
-
-    #[test]
-    fn cancel_probe_returning_false_runs_to_completion() {
-        use cartog_db::Database;
-
-        let (_tmp, root) = tiny_python_project();
-        let db = Database::open_memory().unwrap();
-
-        let probe = || false;
-        let result = index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            Some(&probe),
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .expect("non-cancelling probe must not affect normal indexing");
-        assert!(result.files_indexed >= 2);
-    }
-
-    #[test]
-    fn rerun_after_cancellation_completes_normally() {
-        use cartog_db::Database;
-        use std::sync::atomic::{AtomicBool, Ordering};
-
-        let (_tmp, root) = tiny_python_project();
-        let db = Database::open_memory().unwrap();
-
-        let flag = AtomicBool::new(true);
-        let probe = || flag.load(Ordering::SeqCst);
-        let _ = index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            Some(&probe),
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .expect_err("first run cancels");
-
-        // Flip the probe off — second run must complete and produce a real result.
-        flag.store(false, Ordering::SeqCst);
-        let result = index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            Some(&probe),
-            crate::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .expect("re-run after cancellation must succeed");
-        assert!(result.files_indexed >= 2);
-    }
-
-    // ── Secret redaction integration ──
-
-    /// A project whose single function body embeds a GitHub PAT.
-    fn project_with_secret() -> (tempfile::TempDir, PathBuf) {
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(
-            root.join("conf.py"),
-            "def connect():\n    token = \"ghp_abcdefghijklmnopqrstuvwxyz0123456789\"\n    return token\n",
-        )
-        .unwrap();
-        (tmp, root)
-    }
-
-    fn only_content(db: &Database) -> String {
-        let ids = db.all_content_symbol_ids().unwrap();
-        let map = db.get_symbol_contents_batch(&ids).unwrap();
-        map.values()
-            .map(|(c, _)| c.clone())
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-
-    #[test]
-    fn indexing_redacts_secret_in_symbol_content() {
-        let (_tmp, root) = project_with_secret();
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::enabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let content = only_content(&db);
-        assert!(content.contains("[REDACTED_SECRET]"));
-        assert!(!content.contains("ghp_abcdefghijklmnopqrstuvwxyz0123456789"));
-    }
-
-    #[test]
-    fn redaction_disabled_keeps_secret_verbatim() {
-        let (_tmp, root) = project_with_secret();
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let content = only_content(&db);
-        assert!(content.contains("ghp_abcdefghijklmnopqrstuvwxyz0123456789"));
-        assert!(!content.contains("[REDACTED_SECRET]"));
-    }
-
-    #[test]
-    fn redacted_secret_is_not_searchable_in_fts() {
-        let (_tmp, root) = project_with_secret();
-        let db = Database::open_memory().unwrap();
-        index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::enabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let hits = db
-            .fts5_search("\"ghp_abcdefghijklmnopqrstuvwxyz0123456789\"", 10)
-            .unwrap();
-        assert!(
-            hits.is_empty(),
-            "secret must not be searchable after redaction"
-        );
-    }
-
-    #[test]
-    fn sensitive_file_is_never_indexed() {
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(root.join("a.py"), "def f():\n    pass\n").unwrap();
-        // A code-extension file whose name matches the deny-list.
-        std::fs::write(root.join("id_rsa"), "PRIVATE KEY").unwrap();
-        std::fs::write(root.join(".env"), "API_KEY=ghp_xxxxxxxxxxxxxxxxxxxx").unwrap();
-
-        let db = Database::open_memory().unwrap();
-        let r = index_directory(
-            &db,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::enabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        assert_eq!(r.files_indexed, 1, "only a.py indexes");
-        assert!(
-            r.files_redacted_skipped >= 1,
-            "deny-listed files are skipped"
-        );
-    }
-
-    #[test]
-    fn enabling_redaction_on_warm_index_reindexes_and_scrubs() {
-        let (_tmp, root) = project_with_secret();
-        let db = Database::open_memory().unwrap();
-
-        // First index with redaction OFF: secret is stored verbatim.
-        index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert!(only_content(&db).contains("ghp_abcdefghijklmnopqrstuvwxyz0123456789"));
-
-        // Plain re-index (no --force) with redaction ON must promote to a full
-        // re-index via the policy fingerprint and scrub the stored secret.
-        let r = index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            RedactionConfig::enabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert!(r.redaction_backfilled, "policy change must flag a backfill");
-        let content = only_content(&db);
-        assert!(content.contains("[REDACTED_SECRET]"));
-        assert!(!content.contains("ghp_abcdefghijklmnopqrstuvwxyz0123456789"));
-    }
-
-    #[test]
-    fn content_hash_is_identical_with_redaction_on_vs_off() {
-        let (_tmp, root) = project_with_secret();
-
-        let db_on = Database::open_memory().unwrap();
-        index_directory(
-            &db_on,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::enabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        let db_off = Database::open_memory().unwrap();
-        index_directory(
-            &db_off,
-            &root,
-            true,
-            false,
-            None,
-            None,
-            RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &crate::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-
-        // Hashing keys off raw source, so redaction must not perturb identity.
-        let mut ids_on = db_on.all_content_symbol_ids().unwrap();
-        let mut ids_off = db_off.all_content_symbol_ids().unwrap();
-        ids_on.sort();
-        ids_off.sort();
-        for id in &ids_on {
-            let h_on = db_on.get_symbol(id).unwrap().unwrap().content_hash;
-            let h_off = db_off.get_symbol(id).unwrap().unwrap().content_hash;
-            assert_eq!(h_on, h_off, "content_hash must not depend on redaction");
-        }
-    }
-}
+mod tests;
